General cleanups while working on roll_up
