+++ src/gadgets/merkle_tree.cpp
@@ -1,111 +1,71 @@
 #include "ethsnarks.hpp"
-
+#include "gadgets/merkle_tree.hpp"
 #include "utils.hpp"
 
 namespace ethsnarks {
 
-/**
-* Depending on the address bit, output the correct left/right inputs
-* for the merkle path authentication hash
-*
-* 0 = left
-* 1 = right
-*
-* There are two variables which make up each element of the path,
-* the `input` and the `pathvar`, the input is the leaf or the
-* output from the last hash, and the path var is part of the merkle
-* tree path.
-*
-* The `is_right` parameter decides if the `input` is on the left or
-* right of the hash. These are decided in-circuit using the following
-* method:
-*
-* Left:
-*  (is_right * input) + ((1 - is_right) * pathvar)
-*
-* Right:
-*  ((1 - is_right) * input) + (is_right * pathvar)
-*
-* Each component is split into a & b sides, then added together
-* so the correct variable ends up in the right or left hand side.
-*/
-class merkle_path_selector : public GadgetT
+
+merkle_path_selector::merkle_path_selector(
+    ProtoboardT &in_pb,
+    const VariableT in_input,
+    const VariableT in_pathvar,
+    const VariableT in_is_right,
+    const std::string &in_annotation_prefix
+) :
+    GadgetT(in_pb, in_annotation_prefix),
+    m_input(in_input),
+    m_pathvar(in_pathvar),
+    m_is_right(in_is_right)
 {
-public:
-    const VariableT m_input;
-    const VariableT m_pathvar;
-    const VariableT m_is_right;
+    m_left_a.allocate(in_pb, FMT(this->annotation_prefix, ".left_a"));
+    m_left_b.allocate(in_pb, FMT(this->annotation_prefix, ".left_b"));
+    m_left.allocate(in_pb, FMT(this->annotation_prefix, ".left"));
 
-    VariableT m_left_a;
-    VariableT m_left_b;
-    VariableT m_left;
+    m_right_a.allocate(in_pb, FMT(this->annotation_prefix, ".right_a"));
+    m_right_b.allocate(in_pb, FMT(this->annotation_prefix, ".right_b"));
+    m_right.allocate(in_pb, FMT(this->annotation_prefix, ".right"));
+}
 
-    VariableT m_right_a;
-    VariableT m_right_b;
-    VariableT m_right;
+void merkle_path_selector::generate_r1cs_constraints()
+{
+    this->pb.add_r1cs_constraint(
+        ConstraintT(1 - m_is_right, m_input, m_left_a),
+        FMT(this->annotation_prefix, "1-is_right * input = left_a"));
 
-    merkle_path_selector(
-        ProtoboardT &in_pb,
-        const VariableT in_input,
-        const VariableT in_pathvar,
-        const VariableT in_is_right,
-        const std::string &in_annotation_prefix=""
-    ) :
-        GadgetT(in_pb, in_annotation_prefix),
-        m_input(in_input),
-        m_pathvar(in_pathvar),
-        m_is_right(in_is_right)
-    {
-        m_left_a.allocate(in_pb, FMT(this->annotation_prefix, ".left_a"));
-        m_left_b.allocate(in_pb, FMT(this->annotation_prefix, ".left_b"));
-        m_left.allocate(in_pb, FMT(this->annotation_prefix, ".left"));
+    this->pb.add_r1cs_constraint(ConstraintT(m_is_right, m_pathvar, m_left_b),
+        FMT(this->annotation_prefix, "is_right * pathvar = left_b"));
 
-        m_right_a.allocate(in_pb, FMT(this->annotation_prefix, ".right_a"));
-        m_right_b.allocate(in_pb, FMT(this->annotation_prefix, ".right_b"));
-        m_right.allocate(in_pb, FMT(this->annotation_prefix, ".right"));
-    }
+    this->pb.add_r1cs_constraint(ConstraintT(1, m_left_a + m_left_b, m_left),
+        FMT(this->annotation_prefix, "1 * left_a + left_b = left"));
 
-    void generate_r1cs_constraints()
-    {
-        this->pb.add_r1cs_constraint(
-            ConstraintT(1 - m_is_right, m_input, m_left_a),
-            FMT(this->annotation_prefix, "1-is_right * input = left_a"));
+    this->pb.add_r1cs_constraint(ConstraintT(m_is_right, m_input, m_right_a),
+        FMT(this->annotation_prefix, "is_right * input = right_a"));
 
-        this->pb.add_r1cs_constraint(ConstraintT(m_is_right, m_pathvar, m_left_b),
-            FMT(this->annotation_prefix, "is_right * pathvar = left_b"));
+    this->pb.add_r1cs_constraint(ConstraintT(1 - m_is_right, m_pathvar, m_right_b),
+        FMT(this->annotation_prefix, "1-is_right * pathvar = right_b"));
 
-        this->pb.add_r1cs_constraint(ConstraintT(1, m_left_a + m_left_b, m_left),
-            FMT(this->annotation_prefix, "1 * left_a + left_b = left"));
+    this->pb.add_r1cs_constraint(ConstraintT(1, m_right_a + m_right_b, m_right),
+        FMT(this->annotation_prefix, "1 * right_a + right_b = right"));
+}
 
-        this->pb.add_r1cs_constraint(ConstraintT(m_is_right, m_input, m_right_a),
-            FMT(this->annotation_prefix, "is_right * input = right_a"));
+void merkle_path_selector::generate_r1cs_witness()
+{
+    this->pb.val(m_left_a) = (FieldT::one() - this->pb.val(m_is_right)) * this->pb.val(m_input);
+    this->pb.val(m_left_b) = this->pb.val(m_is_right) * this->pb.val(m_pathvar);
+    this->pb.val(m_left) = this->pb.val(m_left_a) + this->pb.val(m_left_b);
 
-        this->pb.add_r1cs_constraint(ConstraintT(1 - m_is_right, m_pathvar, m_right_b),
-            FMT(this->annotation_prefix, "1-is_right * pathvar = right_b"));
+    this->pb.val(m_right_a) = this->pb.val(m_is_right) * this->pb.val(m_input);
+    this->pb.val(m_right_b) = (FieldT::one() - this->pb.val(m_is_right)) * this->pb.val(m_pathvar);
+    this->pb.val(m_right) = this->pb.val(m_right_a) + this->pb.val(m_right_b);
+}
 
-        this->pb.add_r1cs_constraint(ConstraintT(1, m_right_a + m_right_b, m_right),
-            FMT(this->annotation_prefix, "1 * right_a + right_b = right"));
-    }
+const VariableT& merkle_path_selector::left() const {
+    return m_left;
+}
 
-    void generate_r1cs_witness()
-    {
-        this->pb.val(m_left_a) = (FieldT::one() - this->pb.val(m_is_right)) * this->pb.val(m_input);
-        this->pb.val(m_left_b) = this->pb.val(m_is_right) * this->pb.val(m_pathvar);
-        this->pb.val(m_left) = this->pb.val(m_left_a) + this->pb.val(m_left_b);
-
-        this->pb.val(m_right_a) = this->pb.val(m_is_right) * this->pb.val(m_input);
-        this->pb.val(m_right_b) = (FieldT::one() - this->pb.val(m_is_right)) * this->pb.val(m_pathvar);
-        this->pb.val(m_right) = this->pb.val(m_right_a) + this->pb.val(m_right_b);
-    }
-
-    const VariableT& left() const {
-        return m_left;
-    }
-
-    const VariableT& right() const {
-        return m_right;
-    }
-};
+const VariableT& merkle_path_selector::right() const {
+    return m_right;
+}
 
 
 const VariableArrayT merkle_tree_IVs (ProtoboardT &in_pb)
@@ -147,101 +107,6 @@ const VariableArrayT merkle_tree_IVs (ProtoboardT &in_pb)
     return x;
 }
 
-/**
-* Merkle path authenticator
-*/
-template<typename HashT>
-class merkle_path_authenticator : public GadgetT
-{
-public:
-    const size_t m_depth;
-    const VariableArrayT m_address_bits;
-    const VariableT m_leaf;
-    const VariableT m_expected_root;
-    const VariableArrayT m_path;
-
-    std::vector<merkle_path_selector> m_selectors;
-    std::vector<HashT> m_hashers;
-
-    merkle_path_authenticator(
-        ProtoboardT &in_pb,
-        const size_t in_depth,
-        const VariableArrayT in_address_bits,
-        const VariableArrayT in_IVs,
-        const VariableT in_leaf,
-        const VariableT in_expected_root,
-        const VariableArrayT in_path,
-        const std::string &in_annotation_prefix = ""
-    ) :
-        GadgetT(in_pb, FMT(in_annotation_prefix, " merkle_path_authenticator")),
-        m_depth(in_depth),
-        m_address_bits(in_address_bits),
-        m_leaf(in_leaf),
-        m_expected_root(in_expected_root),
-        m_path(in_path)
-    {
-        assert( in_depth > 0 );
-        assert( in_address_bits.size() == in_depth );
-        assert( in_IVs.size() >= in_depth );
-
-        for( size_t i = 0; i < m_depth; i++ )
-        {
-            if( i == 0 )
-            {
-                m_selectors.push_back(
-                    merkle_path_selector(
-                        in_pb, in_leaf, in_path[i], in_address_bits[i],
-                        FMT(this->annotation_prefix, ".selector_%zu", i)));
-            }
-            else {
-                m_selectors.push_back(
-                    merkle_path_selector(
-                        in_pb, m_hashers[i - 1].result(), in_path[i], in_address_bits[i],
-                        FMT(this->annotation_prefix, ".selector_%zu", i)));
-            }
-
-            m_hashers.push_back(HashT(
-                in_pb, in_IVs[i],
-                {m_selectors[i].left(), m_selectors[i].right()},
-                FMT(this->annotation_prefix, " hasher_%zu", i)));
-        }
-    }
-
-    const VariableT calculated_root() const
-    {
-        return m_hashers[ m_hashers.size() - 1 ].result();
-    }
-
-    bool is_valid() const
-    {
-        return this->pb.val(calculated_root()) == this->pb.val(m_expected_root);
-    }
-
-    void generate_r1cs_constraints()
-    {
-        size_t i;
-        for( i = 0; i < m_hashers.size(); i++ )
-        {
-            m_selectors[i].generate_r1cs_constraints();
-            m_hashers[i].generate_r1cs_constraints();
-        }
-
-        // Ensure root matches calculated path hash
-        this->pb.add_r1cs_constraint(
-            ConstraintT(1, m_hashers[i-1].result(), m_expected_root),
-            FMT(this->annotation_prefix, "expected_root matches"));
-    }
-
-    void generate_r1cs_witness()
-    {
-        size_t i;
-        for( i = 0; i < m_hashers.size(); i++ )
-        {
-            m_selectors[i].generate_r1cs_witness();
-            m_hashers[i].generate_r1cs_witness();
-        }
-    }
-};
 
 // ethsnarks
 }
